# Conflicts:
#	src/melee/cm/camera.c
#	src/melee/ft/chara/ftCommon/ftCo_0A01.c
#	src/melee/ft/chara/ftCommon/ftCo_Bury.c
#	src/melee/gm/gm_1601.c
#	src/melee/gm/gm_1832.c
#	src/melee/gm/gm_1BA8.c
#	src/melee/gm/gmmain_lib.c
#	src/melee/gm/gmregclear.c
#	src/melee/gm/gmregenddisp.c
#	src/melee/gm/gmresult.c
#	src/melee/gm/gmstaffroll.c
#	src/melee/gm/gmtou_0.c
#	src/melee/gm/gmtou_1.c
#	src/melee/gm/gmtoulib.c
#	src/melee/gr/grbigblueroute.c
#	src/melee/gr/grcorneria.c
#	src/melee/gr/groldkongo.c
#	src/melee/gr/grshrineroute.c
#	src/melee/lb/lbaudio_ax.c
#	src/melee/lb/lbheap.c
#	src/melee/mn/mndiagram.c
#	src/melee/mn/mnnamenew.c
#	src/melee/mn/mnvibration.c
#	src/melee/mp/mplib.c
#	src/melee/ty/toy.c
#	src/melee/ty/tydisplay.c
#	src/melee/ty/tyfigupon.c
#	src/sysdolphin/baselib/hsd_3AA7.c
#	src/sysdolphin/baselib/hsd_3B34.c
#	src/sysdolphin/baselib/particle.c

+++ src/melee/gm/gmcamera.c
@@ -343,7 +343,7 @@ void gmCamera_801A28AC(void)
         void* temp_r4 = cmSnap_80031618();
         gmCamera_80479BC8.gcus.x1C = temp_r4;
         if (temp_r4) {
-            if (lbSnap_8001DC0C((int) gmCamera_80479BC8.gcus.x1C) != 0) {
+            if (lbSnap_8001DC0C(gmCamera_80479BC8.gcus.x1C) != 0) {
                 gmCamera_80479BC8.gcus.x20 = lbSnap_8001DF20();
                 gmCamera_801A3048(2);
             } else {
